Prepare for the update feature
This commit updates the overcover repository with the basic go skeleton
I've been using, in preparation for splitting the code up.  The end goal
is to allow overcover to load its threshold from a configuration file in
the current directory, utilizing Viper, and to even allow it to update the
threshold and rewrite that configuration file, allowing the coverage
target to automatically increment.  This will also come with a
comprehensive set of unit tests, which ironically are currently missing.