Add extra dependencies
This commit adds the sources to the dependencies for the build, so that
the binaries will be rebuilt if any of the sources change.  It also adds
the test data files (if any) as dependencies for coverage.out, so that the
tests will be rerun by "make cover" if any of the test data changes.