Allow threshold updates
This commit allows overcover to automatically update a configuration file
with an updated threshold under an algorithm controlled by the
configuration values "min_headroom" and "max_headroom": if the coverage
exceeds the threshold plus max_headroom, the threshold will be updated to
the coverage minus the min_headroom.  A --readonly flag can be given to
report an error instead of updating the configuration.