Enhance coverage computation
This commit enhances overcover by adding support to allow it to directly
count the number of statements in every file in every package in a source
tree.  This is necessary because the "go test" command does not include
counts for files in packages unless the package has at least one
*_test.go.  This support is enabled by simply passing a package
specification on the command line; the package specification should be
the same as provided to "go test".

In addition, this commit also adds the capability to report per-package
coverage summaries and detailed, per-file coverage data.