Makefile updates
This commit contains several updates to the Makefile.  This includes
adding a "tidy" target to run "go mod tidy"; setting up an appropriate
cover-test target that will work for all go projects; reworking the code
that constructs the targets for building binaries; adding code that keeps
.gitignore synchronized; and adding a "help" target.  This commit also
contains an update to .travis.yml and a scripts/overcover.mk which tailors
this Makefile for the overcover repository.